requirements.txt
keras==2.7.0
matplotlib==3.5.0
numpy==1.21.4
opencv_python==4.5.4.60
Pillow==8.4.0
tensorflow==2.7.0
